% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-stt_plot.R
\name{create_plot}
\alias{create_plot}
\title{This function creates a plot combining all plots under 16 scenarios}
\usage{
create_plot(path, show_legend)
}
\arguments{
\item{path}{Path that has the data}

\item{show_legend}{Show the legend or not}
}
\value{
A 16 combined plot
}
\description{
This function creates a plot combining all plots under 16 scenarios
}
