% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-stt_plot.R
\name{crt_stt_plot}
\alias{crt_stt_plot}
\title{This function is doing stt plot for only one scenario, e.g. high_cura}
\usage{
crt_stt_plot(
  data,
  x_var,
  y_var,
  q25_var,
  q75_var,
  color_var = NULL,
  fill_var = NULL,
  show_legend = NULL
)
}
\arguments{
\item{data}{A list contains all median, q25, q75 information}

\item{x_var}{X-axis}

\item{y_var}{Y-axis}

\item{q25_var}{Quantile 25}

\item{q75_var}{Quantile 75}

\item{color_var}{Values for colors}

\item{fill_var}{Values for fill}

\item{show_legend}{Show legend or not}
}
\value{
A ggplot
}
\description{
This function is doing stt plot for only one scenario, e.g. high_cura
}
