% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get_q25q75_median.R
\name{get_all_lists}
\alias{get_all_lists}
\title{This function collects stt data (median, q25, q75) across different levels of mutualism and save the data}
\usage{
get_all_lists(data_files, path, total_time, sample_freq, M0, verbose)
}
\arguments{
\item{data_files}{Data need to be handled}

\item{path}{Path storing data}

\item{total_time}{Island age (simulation time)}

\item{sample_freq}{For formatting}

\item{M0}{Mainland matrix}

\item{verbose}{}
}
\value{
A list listing all median, q25, q75 infomattion
}
\description{
This function collects stt data (median, q25, q75) across different levels of mutualism and save the data
}
