% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get_endemism_list.R
\name{get_endemism_list}
\alias{get_endemism_list}
\title{Get the median, 25th percentile (Q0.25), and 75th percentile (Q0.75) data list for either plant or animal data or
all (plant + animal)}
\usage{
get_endemism_list(cladeinfo, species_type, sample_freq)
}
\arguments{
\item{cladeinfo}{List. A list containing clade information data for multiple islands.
It should include data for both "several_islands_plant" and "several_islands_animal"}

\item{species_type}{Character. Specifies the type of species data to process, either \code{"plant"}
or \code{"animal"}. Based on this argument, the function will select the appropriate dataset
and columns for analysis.}

\item{sample_freq}{Numeric. Defines the frequency of sampling for the simulation.}
}
\value{
A list.
}
\description{
This function processes clade information to calculate the median and quantile statistics
for non-endemic, endemic, and total species counts over a specified sampling frequency.
The function supports both plant and animal data, adjusting automatically based on
the specified species type.
}
