% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get_q25q75_median.R
\name{get_list}
\alias{get_list}
\title{This function gets the edian q25 q75 list of one data set, e.g. data only for none-mutualism case}
\usage{
get_list(island_replicates, total_time, sample_freq, M0, verbose)
}
\arguments{
\item{island_replicates}{"Raw" simulation output from \code{sim_core_mutualism}, not being fomatted yet}

\item{total_time}{island age (simulation time)}

\item{sample_freq}{For fomatting}

\item{M0}{Mainland mutualistic matrix}

\item{verbose}{}
}
\value{
A list
}
\description{
This function gets the edian q25 q75 list of one data set, e.g. data only for none-mutualism case
}
