% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-get_sr_info.R
\name{get_sr_info}
\alias{get_sr_info}
\title{This function summarises the number of plant, animal, and endeminsm on the island}
\usage{
get_sr_info(dataset_name)
}
\arguments{
\item{dataset_name}{Data name}
}
\value{
A list
}
\description{
This function summarises the number of plant, animal, and endeminsm on the island
}
