% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-check_event_freq.R
\name{plot_event_freq}
\alias{plot_event_freq}
\title{This function visualizes the events frequency across all simulations}
\usage{
plot_event_freq(the_data)
}
\arguments{
\item{the_data}{Data being handled}
}
\value{
A ggplot
}
\description{
This function visualizes the events frequency across all simulations
}
